+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Переход с библиотеки `@ntegral/nestjs-sentry` на `@sentry/nestjs`
 - Добавлена проверка сложности пароля в `PasswordField` ([#223](https://gitlab.kozhindev.com/steroids/steroids-nest/-/issues/223))
 - Добавлен параметр `isListenLocalhost` в `RestApplication` для прослушивания только localhost'а, env-переменная для него - `APP_LISTEN_LOCALHOST` ([#215](https://gitlab.kozhindev.com/steroids/steroids-nest/-/issues/215))